FISH-10074 Set relation attribute name using GenAI
