+++ starter-generator/src/main/java/fish/payara/starter/application/generator/CRUDAppGenerator.java
@@ -62,12 +62,13 @@ import static fish.payara.starter.application.util.AttributeType.isPrimitive;
 import static fish.payara.starter.application.util.JPAUtil.ALL_RESERVED_KEYWORDS;
 import static fish.payara.starter.application.util.JavaUtil.getIntrospectionPrefix;
 import static fish.payara.starter.application.util.JavaUtil.getMethodName;
-import static fish.payara.starter.application.util.StringHelper.firstLower;
-import static fish.payara.starter.application.util.StringHelper.firstUpper;
-import static fish.payara.starter.application.util.StringHelper.kebabCase;
-import static fish.payara.starter.application.util.StringHelper.pluralize;
-import static fish.payara.starter.application.util.StringHelper.startCase;
-import static fish.payara.starter.application.util.StringHelper.titleCase;
+import static fish.payara.starter.application.util.StringUtils.firstLower;
+import static fish.payara.starter.application.util.StringUtils.firstUpper;
+import static fish.payara.starter.application.util.StringUtils.kebabCase;
+import static fish.payara.starter.application.util.StringUtils.pluralize;
+import static fish.payara.starter.application.util.StringUtils.singularize;
+import static fish.payara.starter.application.util.StringUtils.startCase;
+import static fish.payara.starter.application.util.StringUtils.titleCase;
 import freemarker.template.Configuration;
 import freemarker.template.Template;
 import freemarker.template.TemplateException;
@@ -504,65 +505,71 @@ public class CRUDAppGenerator {
         String relationshipType = relationship.getRelationshipType();
         String firstEntity = relationship.getFirstEntityClass();
         String secondEntity = relationship.getSecondEntityClass();
-
+        String firstEntityVar = relationship.getRelationshipVarNameInFirstEntity() != null ? relationship.getRelationshipVarNameInFirstEntity() : secondEntity.toLowerCase();
+        String firstEntityVars = pluralize(firstEntityVar);
+        firstEntityVar = singularize(firstEntityVar);
+        String secondEntityVar = relationship.getRelationshipVarNameInSecondEntity() != null ? relationship.getRelationshipVarNameInSecondEntity() : firstEntity.toLowerCase();;
+        String secondEntityVars = pluralize(secondEntityVar);
+        secondEntityVar = singularize(secondEntityVar);
+        
         if (isFirstEntity) {
             switch (relationshipType) {
                 case "||--||": // Exactly one to exactly one
                 case "||--o|": // Exactly one to zero or one
                 case "|o--||": // Zero or one to exactly one
-                    entity.getAttributes().add(new Attribute(secondEntity.toLowerCase(), false, secondEntity));
-                    sbfunc.append("    public ").append(secondEntity).append(" get").append(secondEntity).append("() {\n");
-                    sbfunc.append("        return ").append(secondEntity.toLowerCase()).append(";\n");
+                    entity.getAttributes().add(new Attribute(firstEntityVar, false, secondEntity));
+                    sbfunc.append("    public ").append(secondEntity).append(" get").append(firstUpper(firstEntityVar)).append("() {\n");
+                    sbfunc.append("        return ").append(firstEntityVar).append(";\n");
                     sbfunc.append("    }\n\n");
-                    sbfunc.append("    public void set").append(secondEntity).append("(").append(secondEntity).append(" ").append(secondEntity.toLowerCase()).append(") {\n");
-                    sbfunc.append("        this.").append(secondEntity.toLowerCase()).append(" = ").append(secondEntity.toLowerCase()).append(";\n");
+                    sbfunc.append("    public void set").append(firstUpper(firstEntityVar)).append("(").append(secondEntity).append(" ").append(firstEntityVar).append(") {\n");
+                    sbfunc.append("        this.").append(firstEntityVar).append(" = ").append(firstEntityVar).append(";\n");
                     sbfunc.append("    }\n\n");
-                    sb.append("    @OneToOne(mappedBy = \"").append(firstEntity.toLowerCase()).append("\")\n");
-                    sb.append("    private ").append(secondEntity).append(" ").append(secondEntity.toLowerCase()).append(";\n");
+                    sb.append("    @OneToOne(mappedBy = \"").append(secondEntityVar).append("\")\n");
+                    sb.append("    private ").append(secondEntity).append(" ").append(firstEntityVar).append(";\n");
                     break;
                 case "||--|{": // Exactly one to one or more
                 case "||--o{": // Exactly one to zero or more
-//                    entity.getAttributes().add(new Attribute(secondEntity.toLowerCase(), secondEntityObj, true, relationship.getProperty()));
-                    sbfunc.append("    public List<").append(secondEntity).append("> get").append(pluralize(secondEntity)).append("() {\n");
-                    sbfunc.append("        return ").append(pluralize(secondEntity.toLowerCase())).append(";\n");
+                    entity.getAttributes().add(new Attribute(firstEntityVars, true, secondEntity));
+                    sbfunc.append("    public List<").append(secondEntity).append("> get").append(firstUpper(firstEntityVars)).append("() {\n");
+                    sbfunc.append("        return ").append(firstEntityVars).append(";\n");
                     sbfunc.append("    }\n\n");
-                    sbfunc.append("    public void set").append(pluralize(secondEntity)).append("(List<").append(secondEntity).append("> ").append(pluralize(secondEntity.toLowerCase())).append(") {\n");
-                    sbfunc.append("        this.").append(pluralize(secondEntity.toLowerCase())).append(" = ").append(pluralize(secondEntity.toLowerCase())).append(";\n");
+                    sbfunc.append("    public void set").append(firstUpper(firstEntityVars)).append("(List<").append(secondEntity).append("> ").append(firstEntityVars).append(") {\n");
+                    sbfunc.append("        this.").append(firstEntityVars).append(" = ").append(firstEntityVars).append(";\n");
                     sbfunc.append("    }\n\n");
                     _imports.add(model.getImportPrefix() + ".json.bind.annotation.JsonbTransient");
                     _imports.add("java.util.List");
                     sb.append("    @JsonbTransient\n");
-                    sb.append("    @OneToMany(mappedBy = \"").append(firstEntity.toLowerCase()).append("\")\n");
-                    sb.append("    private List<").append(secondEntity).append("> ").append(pluralize(secondEntity.toLowerCase())).append(";\n");
+                    sb.append("    @OneToMany(mappedBy = \"").append(secondEntityVar).append("\")\n");
+                    sb.append("    private List<").append(secondEntity).append("> ").append(firstEntityVars).append(";\n");
                     break;
                 case "}|--||": // One or more to exactly one
                 case "}o--||": // Zero or more to exactly one
-                    entity.getAttributes().add(new Attribute(secondEntity.toLowerCase(), false, secondEntity));
-                    sbfunc.append("    public ").append(secondEntity).append(" get").append(secondEntity).append("() {\n");
-                    sbfunc.append("        return ").append(secondEntity.toLowerCase()).append(";\n");
+                    entity.getAttributes().add(new Attribute(firstEntityVar, false, secondEntity));
+                    sbfunc.append("    public ").append(secondEntity).append(" get").append(firstUpper(firstEntityVar)).append("() {\n");
+                    sbfunc.append("        return ").append(firstEntityVar).append(";\n");
                     sbfunc.append("    }\n\n");
-                    sbfunc.append("    public void set").append(secondEntity).append("(").append(secondEntity).append(" ").append(secondEntity.toLowerCase()).append(") {\n");
-                    sbfunc.append("        this.").append(secondEntity.toLowerCase()).append(" = ").append(secondEntity.toLowerCase()).append(";\n");
+                    sbfunc.append("    public void set").append(firstUpper(firstEntityVar)).append("(").append(secondEntity).append(" ").append(firstEntityVar).append(") {\n");
+                    sbfunc.append("        this.").append(firstEntityVar).append(" = ").append(firstEntityVar).append(";\n");
                     sbfunc.append("    }\n\n");
                     sb.append("    @ManyToOne\n");
-                    sb.append("    private ").append(secondEntity).append(" ").append(secondEntity.toLowerCase()).append(";\n");
+                    sb.append("    private ").append(secondEntity).append(" ").append(firstEntityVar).append(";\n");
                     break;
                 case "}o--o{": // Zero or more to zero or more
                 case "}|--o{": // One or more to zero or more
                 case "}o--|{": // Zero or more to one or more
                 case "}|--|{": // One or more to one or more
-//                    entity.getAttributes().add(new Attribute(secondEntity.toLowerCase(), secondEntityObj, true, relationship.getProperty()));
-                    sbfunc.append("    public List<").append(secondEntity).append("> get").append(pluralize(secondEntity)).append("() {\n");
-                    sbfunc.append("        return ").append(pluralize(secondEntity.toLowerCase())).append(";\n");
+                    entity.getAttributes().add(new Attribute(firstEntityVars, true, secondEntity));
+                    sbfunc.append("    public List<").append(secondEntity).append("> get").append(firstUpper(firstEntityVars)).append("() {\n");
+                    sbfunc.append("        return ").append(firstEntityVars).append(";\n");
                     sbfunc.append("    }\n\n");
-                    sbfunc.append("    public void set").append(pluralize(secondEntity)).append("(List<").append(secondEntity).append("> ").append(pluralize(secondEntity.toLowerCase())).append(") {\n");
-                    sbfunc.append("        this.").append(pluralize(secondEntity.toLowerCase())).append(" = ").append(pluralize(secondEntity.toLowerCase())).append(";\n");
+                    sbfunc.append("    public void set").append(firstUpper(firstEntityVars)).append("(List<").append(secondEntity).append("> ").append(firstEntityVars).append(") {\n");
+                    sbfunc.append("        this.").append(firstEntityVars).append(" = ").append(firstEntityVars).append(";\n");
                     sbfunc.append("    }\n\n");
                     _imports.add(model.getImportPrefix() + ".json.bind.annotation.JsonbTransient");
                     _imports.add("java.util.List");
                     sb.append("    @JsonbTransient\n");
-                    sb.append("    @ManyToMany(mappedBy = \"").append(firstEntity.toLowerCase()).append("\")\n");
-                    sb.append("    private List<").append(secondEntity).append("> ").append(pluralize(secondEntity.toLowerCase())).append(";\n");
+                    sb.append("    @ManyToMany(mappedBy = \"").append(secondEntityVar).append("\")\n");
+                    sb.append("    private List<").append(secondEntity).append("> ").append(firstEntityVars).append(";\n");
                     break;
             }
         } else {
@@ -570,55 +577,55 @@ public class CRUDAppGenerator {
                 case "||--||": // Exactly one to exactly one
                 case "||--o|": // Exactly one to zero or one
                 case "|o--||": // Zero or one to exactly one
-//                    entity.getAttributes().add(new Attribute(firstEntity.toLowerCase(), firstEntityObj, false, relationship.getProperty()));
-                    sbfunc.append("    public ").append(firstEntity).append(" get").append(firstEntity).append("() {\n");
-                    sbfunc.append("        return ").append(firstEntity.toLowerCase()).append(";\n");
+                    entity.getAttributes().add(new Attribute(secondEntityVar, false, secondEntity));
+                    sbfunc.append("    public ").append(firstEntity).append(" get").append(firstUpper(secondEntityVar)).append("() {\n");
+                    sbfunc.append("        return ").append(secondEntityVar).append(";\n");
                     sbfunc.append("    }\n\n");
-                    sbfunc.append("    public void set").append(firstEntity).append("(").append(firstEntity).append(" ").append(firstEntity.toLowerCase()).append(") {\n");
-                    sbfunc.append("        this.").append(firstEntity.toLowerCase()).append(" = ").append(firstEntity.toLowerCase()).append(";\n");
+                    sbfunc.append("    public void set").append(firstUpper(secondEntityVar)).append("(").append(firstEntity).append(" ").append(secondEntityVar).append(") {\n");
+                    sbfunc.append("        this.").append(secondEntityVar).append(" = ").append(secondEntityVar).append(";\n");
                     sbfunc.append("    }\n\n");
                     _imports.add(model.getImportPrefix() + ".json.bind.annotation.JsonbTransient");
                     sb.append("    @JsonbTransient\n");
                     sb.append("    @OneToOne\n");
-                    sb.append("    @JoinColumn(name = \"").append(firstEntity.toLowerCase()).append("_id\")\n");
-                    sb.append("    private ").append(firstEntity).append(" ").append(firstEntity.toLowerCase()).append(";\n");
+                    sb.append("    @JoinColumn(name = \"").append(secondEntityVar).append("_id\")\n");
+                    sb.append("    private ").append(firstEntity).append(" ").append(secondEntityVar).append(";\n");
                     break;
                 case "||--|{": // Exactly one to one or more
                 case "||--o{": // Exactly one to zero or more
-                    entity.getAttributes().add(new Attribute(firstEntity.toLowerCase(), false, firstEntity));
-                    sbfunc.append("    public ").append(firstEntity).append(" get").append(firstEntity).append("() {\n");
-                    sbfunc.append("        return ").append(firstEntity.toLowerCase()).append(";\n");
+                    entity.getAttributes().add(new Attribute(secondEntityVar, false, firstEntity));
+                    sbfunc.append("    public ").append(firstEntity).append(" get").append(firstUpper(secondEntityVar)).append("() {\n");
+                    sbfunc.append("        return ").append(secondEntityVar).append(";\n");
                     sbfunc.append("    }\n\n");
-                    sbfunc.append("    public void set").append(firstEntity).append("(").append(firstEntity).append(" ").append(firstEntity.toLowerCase()).append(") {\n");
-                    sbfunc.append("        this.").append(firstEntity.toLowerCase()).append(" = ").append(firstEntity.toLowerCase()).append(";\n");
+                    sbfunc.append("    public void set").append(firstUpper(secondEntityVar)).append("(").append(firstEntity).append(" ").append(secondEntityVar).append(") {\n");
+                    sbfunc.append("        this.").append(secondEntityVar).append(" = ").append(secondEntityVar).append(";\n");
                     sbfunc.append("    }\n\n");
                     sb.append("    @ManyToOne\n");
-                    sb.append("    @JoinColumn(name = \"").append(firstEntity.toLowerCase()).append("_id\")\n");
-                    sb.append("    private ").append(firstEntity).append(" ").append(firstEntity.toLowerCase()).append(";\n");
+                    sb.append("    @JoinColumn(name = \"").append(secondEntityVar).append("_id\")\n");
+                    sb.append("    private ").append(firstEntity).append(" ").append(secondEntityVar).append(";\n");
                     break;
 //                case "}|--||": // One or more to exactly one
 //                case "}o--||": // Zero or more to exactly one
 //                    sb.append("    @OneToMany\n");
-//                    sb.append("    @JoinColumn(name = \"").append(firstEntity.toLowerCase()).append("_id\")\n");
-//                    sb.append("    private ").append(firstEntity).append(" ").append(firstEntity.toLowerCase()).append(";\n");
+//                    sb.append("    @JoinColumn(name = \"").append(secondEntityVar).append("_id\")\n");
+//                    sb.append("    private ").append(firstEntity).append(" ").append(secondEntityVar).append(";\n");
 //                    break;
                 case "}o--o{": // Zero or more to zero or more
                 case "}|--o{": // One or more to zero or more
                 case "}o--|{": // Zero or more to one or more
                 case "}|--|{": // One or more to one or more
-//                    entity.getAttributes().add(new Attribute(firstEntity.toLowerCase(), firstEntityObj, true, relationship.getProperty()));
-                    sbfunc.append("    public List<").append(firstEntity).append("> get").append(pluralize(firstEntity)).append("() {\n");
-                    sbfunc.append("        return ").append(pluralize(firstEntity.toLowerCase())).append(";\n");
+                    entity.getAttributes().add(new Attribute(secondEntityVars, true, firstEntity));
+                    sbfunc.append("    public List<").append(firstEntity).append("> get").append(firstUpper(secondEntityVars)).append("() {\n");
+                    sbfunc.append("        return ").append(secondEntityVars).append(";\n");
                     sbfunc.append("    }\n\n");
-                    sbfunc.append("    public void set").append(pluralize(firstEntity)).append("(List<").append(firstEntity).append("> ").append(pluralize(firstEntity.toLowerCase())).append(") {\n");
-                    sbfunc.append("        this.").append(pluralize(firstEntity.toLowerCase())).append(" = ").append(pluralize(firstEntity.toLowerCase())).append(";\n");
+                    sbfunc.append("    public void set").append(firstUpper(secondEntityVars)).append("(List<").append(firstEntity).append("> ").append(secondEntityVars).append(") {\n");
+                    sbfunc.append("        this.").append(secondEntityVars).append(" = ").append(secondEntityVars).append(";\n");
                     sbfunc.append("    }\n\n");
                     _imports.add(model.getImportPrefix() + ".json.bind.annotation.JsonbTransient");
                     _imports.add("java.util.List");
                     sb.append("    @JsonbTransient\n");
                     sb.append("    @ManyToMany\n");
-                    sb.append("    @JoinColumn(name = \"").append(firstEntity.toLowerCase()).append("_id\")\n");
-                    sb.append("    private List<").append(firstEntity).append("> ").append(pluralize(firstEntity.toLowerCase())).append(";\n");
+                    sb.append("    @JoinColumn(name = \"").append(secondEntityVar).append("_id\")\n");
+                    sb.append("    private List<").append(firstEntity).append("> ").append(secondEntityVars).append(";\n");
                     break;
             }
         }
